requirements.txt
line-bot-sdk==3.5.0
flask==2.3.3
python-dotenv==1.0.0 